- Fixed bug #13246: ezorderitem->VATValue needs to float rather then int.
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/trunk@22569 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/classes/ezorderitem.php
@@ -70,7 +70,7 @@ class eZOrderItem extends eZPersistentObject
                                                            'default' => 0,
                                                            'required' => true ),
                                          'vat_value' => array( 'name' => 'VATValue',
-                                                               'datatype' => 'integer',
+                                                               'datatype' => 'float',
                                                                'default' => 0,
                                                                'required' => true ),
                                          'is_vat_inc' => array( 'name' => 'IsVATIncluded',
